# src/flask/openapisample.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Full Stack FastAPI Project",
    "description": "Sample API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/scheduler/time-off-requests": {
      "get": {
        "tags": [
          "scheduler"
        ],
        "summary": "Get My Time Off",
        "description": "Get current user's time-off requests (member:limited scope required).",
        "operationId": "get_my_time_off_api_v1_scheduler_time_off_requests_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ISO date YYYY-MM-DD",
              "title": "Start Date"
            },
            "description": "ISO date YYYY-MM-DD"
          },
          {
            "name": "end_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ISO date YYYY-MM-DD",
              "title": "End Date"
            },
            "description": "ISO date YYYY-MM-DD"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimeOffRequestsPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/scheduler/time-off-request": {
      "post": {
        "tags": [
          "scheduler"
        ],
        "summary": "Create Time Off Request",
        "description": "Create a time-off request for the current user.",
        "operationId": "create_time_off_request_api_v1_scheduler_time_off_request_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TimeOffRequestCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimeOffRequestPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/scheduler/time-off-requests/{time_off_id}/approve": {
      "patch": {
        "tags": [
          "scheduler"
        ],
        "summary": "Approve Time Off Request",
        "description": "Approve a time-off request (admin only).",
        "operationId": "approve_time_off_request_api_v1_scheduler_time_off_requests__time_off_id__approve_patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "time_off_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Time Off Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimeOffRequestPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/scheduler/time-off-requests/{time_off_id}/decline": {
      "patch": {
        "tags": [
          "scheduler"
        ],
        "summary": "Decline Time Off Request",
        "description": "Decline a time-off request (admin only).",
        "operationId": "decline_time_off_request_api_v1_scheduler_time_off_requests__time_off_id__decline_patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "time_off_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Time Off Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimeOffRequestPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/scheduler/time-off-requests/{time_off_id}": {
      "delete": {
        "tags": [
          "scheduler"
        ],
        "summary": "Delete Time Off Request",
        "description": "Delete a time-off request (owner only).",
        "operationId": "delete_time_off_request_api_v1_scheduler_time_off_requests__time_off_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "time_off_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Time Off Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/scheduler/my-assignments": {
      "get": {
        "tags": [
          "scheduler"
        ],
        "summary": "Get My Assignments",
        "description": "Get current user's own assignments (member:limited scope required).",
        "operationId": "get_my_assignments_api_v1_scheduler_my_assignments_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssignmentsPublic"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/scheduler/calendar": {
      "get": {
        "tags": [
          "scheduler"
        ],
        "summary": "Get Calendar Assignments",
        "description": "Get assignments for a date range (member:limited scope required).",
        "operationId": "get_calendar_assignments_api_v1_scheduler_calendar_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "start_date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "ISO date YYYY-MM-DD",
              "title": "Start Date"
            },
            "description": "ISO date YYYY-MM-DD"
          },
          {
            "name": "end_date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "ISO date YYYY-MM-DD",
              "title": "End Date"
            },
            "description": "ISO date YYYY-MM-DD"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssignmentsPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/scheduler/calendar-with-names": {
      "get": {
        "tags": [
          "scheduler"
        ],
        "summary": "Get Calendar With Names",
        "description": "Get assignments for a date range, including user email (member:limited scope required).",
        "operationId": "get_calendar_with_names_api_v1_scheduler_calendar_with_names_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "start_date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "ISO date YYYY-MM-DD",
              "title": "Start Date"
            },
            "description": "ISO date YYYY-MM-DD"
          },
          {
            "name": "end_date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "ISO date YYYY-MM-DD",
              "title": "End Date"
            },
            "description": "ISO date YYYY-MM-DD"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnrichedAssignmentsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/scheduler/my-calendar": {
      "get": {
        "tags": [
          "scheduler"
        ],
        "summary": "Get My Calendar",
        "description": "Get current user's own assignments in a date range (member:limited scope required).",
        "operationId": "get_my_calendar_api_v1_scheduler_my_calendar_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "start_date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "ISO date YYYY-MM-DD",
              "title": "Start Date"
            },
            "description": "ISO date YYYY-MM-DD"
          },
          {
            "name": "end_date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "ISO date YYYY-MM-DD",
              "title": "End Date"
            },
            "description": "ISO date YYYY-MM-DD"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssignmentsPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/scheduler/": {
      "get": {
        "tags": [
          "scheduler"
        ],
        "summary": "List Assignments",
        "description": "List all assignments (scheduler:admin scope required).",
        "operationId": "list_assignments_api_v1_scheduler__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssignmentsPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "scheduler"
        ],
        "summary": "Create Assignment",
        "description": "Create a new assignment (scheduler:admin scope required).",
        "operationId": "create_assignment_api_v1_scheduler__post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignmentCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssignmentPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/scheduler/{assignment_id}": {
      "get": {
        "tags": [
          "scheduler"
        ],
        "summary": "Get Assignment",
        "description": "Get a single assignment (scheduler:admin scope required).",
        "operationId": "get_assignment_api_v1_scheduler__assignment_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "assignment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Assignment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssignmentPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "scheduler"
        ],
        "summary": "Update Assignment",
        "description": "Update an assignment (scheduler:admin scope required).",
        "operationId": "update_assignment_api_v1_scheduler__assignment_id__patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "assignment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Assignment Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignmentUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssignmentPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "scheduler"
        ],
        "summary": "Delete Assignment",
        "description": "Delete an assignment (scheduler:admin scope required).",
        "operationId": "delete_assignment_api_v1_scheduler__assignment_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "assignment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Assignment Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/scheduler/bulk": {
      "post": {
        "tags": [
          "scheduler"
        ],
        "summary": "Bulk Assign",
        "description": "Bulk create assignments for multiple users on the same date\n(scheduler:admin scope required).",
        "operationId": "bulk_assign_api_v1_scheduler_bulk_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkAssignRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkAssignResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TimeOffRequestCreate": {
        "properties": {
          "date": {
            "type": "string",
            "format": "date-time",
            "title": "Date"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 4000
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "date"
        ],
        "title": "TimeOffRequestCreate"
      },
...
      "TimeOffRequestsPublic": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/TimeOffRequestPublic"
            },
            "type": "array",
            "title": "Data"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "data",
          "count"
        ],
        "title": "TimeOffRequestsPublic"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {

            },
            "tokenUrl": "/api/v1/login/access-token"
          }
        }
      }
    }
  }
}